% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwdata-package.R
\docType{package}
\name{cwdata-package}
\alias{cwdata}
\alias{cwdata-package}
\title{cwdata: Communicating with Data}
\description{
A project created during ETC5523 tutorial. Contains data and resources from the lectures
             and tutorials.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/petestylianos/cwdata}
  \item Report bugs at \url{https://github.com/petestylianos/cwdata/issues}
}

}
\author{
\strong{Maintainer}: Panagiotis Stylianos \email{pete.styl@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
