% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crop_yield.R
\name{get_crop_yield}
\alias{get_crop_yield}
\title{Add together two numbers}
\usage{
get_crop_yield()
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
desc
}
\details{
detail
}
\examples{
add(1, 1)
add(10, 1)

}
