% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_crop_yields.R
\docType{data}
\name{key_crop_yields}
\alias{key_crop_yields}
\title{Title}
\format{
A data frame with 53940 rows ans 10 variables;
\describe{
\item{price}{price, in US dollar}
\item{carat}{weight of dimond, in carats}
...

}
}
\source{
where you get the data
}
\usage{
key_crop_yields
}
\description{
More details about the data
}
\section{Data contents}{

}

\examples{
print(key_crop_yields)

}
\keyword{datasets}
