% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{n_countries}
\alias{n_countries}
\title{"counts the number of countries"}
\usage{
n_countries(data)
}
\arguments{
\item{data}{A dataset cotaining a country column}
}
\description{
"counts the number of countries"
}
