% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack-normalise.R
\name{stack_normalise}
\alias{stack_normalise}
\alias{stack_normalize}
\title{Normalise a vector of numbers to sum to 1 (Title)}
\usage{
stack_normalise(x)
}
\arguments{
\item{x}{A vector of numbers (what param is)}
}
\value{
The vector x normalised to proportions
}
\description{
Normalise a vector of numbers to sum to 1 (Title)
}
\examples{
x <- c(10, 30, 40)
stack_normalise(x) 
}
