% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwdata-package.R
\docType{package}
\name{cwdata-package}
\alias{cwdata}
\alias{cwdata-package}
\title{cwdata: Communicating with Data Resources}
\description{
Contains data and examples from the CWD course.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mayunk264/cwdata}
  \item Report bugs at \url{https://github.com/mayunk264/cwdata/issues}
}

}
\author{
\strong{Maintainer}: Mayunk Bharadwaj \email{mbha0014@student.monash.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
