% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwdata-package.R
\docType{package}
\name{cwdata-package}
\alias{cwdata}
\alias{cwdata-package}
\title{cwdata: Communicating with Data Resources}
\description{
Contains data and examples from the CWD cource.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Yiwen-Zhang-259/cwdata.git}
}

}
\author{
\strong{Maintainer}: Yiwen Zhang \email{yzha0633@student.monash.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
