% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crop_yield.R
\name{get_crop_yield}
\alias{get_crop_yield}
\title{Get crop yield}
\usage{
get_crop_yield()
}
\description{
This function gets crop yield.
}
\details{
This function does get crop yield.
}
\examples{
get_export_yield()
}
