% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_crop_yields.R
\docType{data}
\name{key_crop_yields}
\alias{key_crop_yields}
\title{roxygen comment}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 143825 rows and 5 columns.
}
\source{
Tidy Tuesday
}
\usage{
key_crop_yields
}
\description{
this dataset contains key crop yields from the 249 countries.
}
\keyword{datasets}
