% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_crop_yields.R
\docType{data}
\name{key_crop_yields}
\alias{key_crop_yields}
\title{Key Crop Yields Dataset}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{entity}{Country names}
\item{code}{Country code}
...
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
key_crop_yields
}
\description{
Key Crop Yields Dataset
}
\section{Background}{

The data was collected from ... over the course.
}

\keyword{datasets}
