% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_crop_yields.R
\docType{data}
\name{key_crop_yields}
\alias{key_crop_yields}
\title{Title}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{entity}{coutry name}
\item{code}{cpountry code}
\item{year}{year recorded}
\item{crop}{crop produces}
\item{tonnes_per_hectare}{tonnes per hectare produced}

...
}
}
\source{
Where you get the data

\url{https://raw.githubusercontent.com/rfordatascience/tidytuesday/master/data/2020/2020-09-01/key_crop_yields.csv}
}
\usage{
key_crop_yields
}
\description{
More details aout the data
}
\examples{


A dataset containing 143825 observation and 5 variablesdevtools::document()

}
\keyword{datasets}
