% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_crop_yields.R
\docType{data}
\name{key_crop_yields}
\alias{key_crop_yields}
\title{Title of the data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 143825 rows and 5 columns.
}
\source{
Where you get the data
}
\usage{
key_crop_yields
}
\description{
More details about the data
}
\section{Backgroud}{

The data was collected from ... over the course of ...

\describe{
\item{Entity}{Country}
\item{toones_per_hectare}{yield}
...
}
}

\examples{
print(key_crop_yields)
}
\keyword{datasets}
