% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_crop_yields.R
\docType{data}
\name{key_crop_yields}
\alias{key_crop_yields}
\title{key crop yields
tbl_df
tbl
data.frame}
\format{
A data frame with 143825 rows and 5 variables:
\describe{
\item{entity}{country}
\item{code}{3 letter country code}
\item{year}{year from 1961 to 2018}
\item{crop}{the crops arwheat, rice, maize, soybeans, potatoes, beans, peas, cassava, barley, cocoa, and bananas)`}
\item{tonnes_per_hectare}{produce of crop in tonnes per hectare}
...
}
}
\source{
\url{'https://raw.githubusercontent.com/rfordatascience/tidytuesday/master/data/2020/2020-09-01/key_crop_yields.csv'}
}
\usage{
key_crop_yields
}
\description{
More details about the data
}
\section{Data Contents}{

The data was collected from... during...
}

\examples{
head(key_crop_yields)
}
\keyword{datasets}
