% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_crop_yields.R
\docType{data}
\name{key_crop_yields}
\alias{key_crop_yields}
\title{Key crop yields}
\format{
A data frame with 143,825 rows and 5 variables.
\describe{
\item{entity}{Country or region name}
\item{code}{Country code (NA for regions or continents)}
\item{year}{Year}
\item{crop}{Crop type}
\item{yield}{Yield in tonnes_per_hectare}
}
}
\source{
\url{https://raw.githubusercontent.com/rfordatascience/tidytuesday/master/data/2020/2020-09-01/key_crop_yields.csv}
}
\usage{
key_crop_yields
}
\description{
A dataset that contains the agricultural yields across crop types and by
entity from 1960 to 2018.
}
\keyword{datasets}
