% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_crop_yields.R
\docType{data}
\name{key_crop_yields}
\alias{key_crop_yields}
\title{The data was collected from Tidy Tuesday Git Repository}
\format{
A data frame with 53948 rows and 10 variables:
\describe{
\item{Entity}{Country}
\item{tonnes_per_hectare}{yield}

...
}
}
\source{
'https://raw.githubusercontent.com/rfordatascience/tidytuesday/master/data/2020/2020-09-01/key_crop_yields.csv'
}
\usage{
key_crop_yields
}
\description{
The data was collected from Tidy Tuesday Git Repository
}
\examples{
print(key_crop_yields)
}
\keyword{datasets}
