% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_crop_yields.R
\docType{data}
\name{key_crop_yields}
\alias{key_crop_yields}
\title{Title of the data}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}
...
}
}
\source{
where you got the data
}
\usage{
key_crop_yields
}
\description{
More details about the data
}
\section{Backgroud}{

The data was collected from ... over the course of ...
}

\examples{
print(key_crop_yields)
}
\keyword{datasets}
