% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_crop_yields.R
\docType{data}
\name{key_crop_yields}
\alias{key_crop_yields}
\title{Title of the data}
\format{
A data frame with 143,825 rows and 5 variables:
\describe{
\item{entity}{country names}
\item{code}{country code}
\item{year}{year}
\item{crop}{crop type}
\item{tonnes_per_hectare}{yield}
}
}
\source{

}
\usage{
key_crop_yields
}
\description{
More details about the data
}
\section{Background}{

The data was collected from ... over the course of ....
}

\examples{
print(key_crop_yields)
}
\keyword{datasets}
