% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_crop_yields.R
\docType{data}
\name{key_crop_yields}
\alias{key_crop_yields}
\title{Global yield of key crops}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 143825 rows and 5 columns.
}
\source{
Tidy Tuesday
}
\usage{
key_crop_yields
}
\description{
This dataset contains yields of 11 key crops from 249 countries.
}
\keyword{datasets}
