% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_crop_yields.R
\docType{data}
\name{key_crop_yields}
\alias{key_crop_yields}
\title{Title}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}
...
}
}
\source{
Where you got the data
}
\usage{
key_crop_yields
}
\arguments{
\item{x}{}
}
\description{
Title
}
\section{Background}{

The data was colected from ...
}

\examples{
print(key_crop_yields)
}
\keyword{datasets}
