% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{n_countries}
\alias{n_countries}
\title{Count the number of countries in a dataset (title)}
\usage{
n_countries(data)
}
\arguments{
\item{data}{A dataset containing a \code{country} column (input)}
}
\description{
Count the number of countries in a dataset (title)
}
\examples{
n_countries(key_crop_yields)

}
