% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack-normalise.R
\name{stack_normalise}
\alias{stack_normalise}
\title{Normalise numerical vectors}
\usage{
stack_normalise(x)
}
\arguments{
\item{x}{A numerical vector with positive values (i.e. 0 or greater).}
}
\value{
A numerical vector of the same size with proportions.
The sum of the numerical values in the return values should
add up to 1.
}
\description{
This function takes a numerical vector
and computes the proportion. Missing values
will be removed from the input vector but remain in the output vector.
}
\examples{
stack_normalise(c(10, 30, 40))
stack_normalise(c(75, 0, 5, 20, NA))
stack_normalise(c(NA, NA, 10))
}
