% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack-normalise.R
\name{stack_normalise}
\alias{stack_normalise}
\alias{stack_normalize}
\title{Normalise vector of numbers to sum to 1}
\usage{
stack_normalise(x)
}
\arguments{
\item{x}{is a vector of numbers}
}
\value{
the vector x normalised to proportions
}
\description{
Normalise vector of numbers to sum to 1
}
\examples{
x <- c(10, 30, 40)
stack_normalize(x)
c(0.125, 0.375, .5)
x <- c(75, 0, 5, 20, NA)
stack_normalise(x)
x <- c(NA, NA, 10)
stack_normalize(x)
c(NA, NA, 1)

}
