% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack-normalise.R
\name{stack_normalise}
\alias{stack_normalise}
\alias{stack_normalize}
\title{Convert numbers into proportions}
\usage{
stack_normalise(x)

stack_normalize(x)
}
\arguments{
\item{x}{A vector of numbers}
}
\value{

}
\description{
Convert numbers into proportions
}
\examples{

x <- c(10, 30, 40)
stack_normalize(x)
x <- c(75, 0, 5, 20, NA)
stack_normalise(x)
x <- c(NA, NA, 10)
stack_normalize(x)
}
