% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_normalise.R
\name{stack_normalise}
\alias{stack_normalise}
\alias{stack_normalize}
\title{Normalise a vector of numbers to sum to 1}
\usage{
stack_normalise(x)
}
\arguments{
\item{x}{A vector of numbers}
}
\value{
The vector x normalised to proportions
}
\description{
Normalise a vector of numbers to sum to 1
}
\examples{
x <- c(10, 30, 40)
stack_normalize(x)
x <- c(75, 0, 5, 20, NA)
stack_normalise(x)
x <- c(NA, NA, 10)
stack_normalize(x)

}
